+++ frontend/src/components/Card.vue
@@ -1,15 +1,18 @@
 <template>
+        <b-overlay :show="show" rounded="lg">
           <b-card
             :img-src= "require(`../assets/img/cards/${imageId}.jpg`)"
             class='border-0 mx-3'
+            :aria-hidden="show ? 'true' : null"
             img-alt='' img-top no-body
           >
-            <!-- Info -->
-            <b-icon
-              icon='info-circle-fill'
-              class='info-icon position-absolute cursor-pointer'
-              v-b-tooltip="{ title: `${desktop.description ? desktop.description : $t(`components.desktop-cards.no-info-default`)}`, placement: 'top', customClass: 'isard-tooltip', trigger: 'hover' }"
-            ></b-icon>
+            <vue-fab v-if="desktop.type === 'persistent'" mainBtnColor="#bcc6cc" class='info-icon position-absolute' size='small' unfoldDirection='down' :scrollAutoHide="false">
+                <fab-item @clickItem="onClickDeleteDesktop" :idx="0" icon='delete' color='#e34934' titleBgColor='#e6e8eb' v-b-tooltip="{ title: `${$t('components.desktop-cards.actions.delete')}`, placement: 'top', customClass: 'isard-tooltip', trigger: 'hover' }">
+                </fab-item>
+              <!-- <fab-item  :idx="1" title="Edit" icon='mode_edit' color='#c2d5f2' titleBgColor='#e6e8eb'/>
+              <fab-item  :idx="2" title="Template" icon='library_books' titleBgColor='#e6e8eb'/>
+              <fab-item  :idx="3" title="Change image" icon='autorenew' titleBgColor='#e6e8eb'/> -->
+            </vue-fab>
 
             <div class='p-2 h-100 d-flex flex-wrap flex-column' :class="{'startedHighlight': desktopState === desktopStates.started}">
               <div class='flex-grow-1'>
@@ -104,6 +107,7 @@
               </div>
             </div>
           </b-card>
+        </b-overlay>
 </template>
 
 <script>
@@ -142,6 +146,28 @@ export default {
     },
     isWaiting (viewer) {
       return this.getDefaultViewer && (this.waitingIp && DesktopUtils.viewerNeedsIp(viewer))
+    },
+    onClickDeleteDesktop (toast) {
+      this.$snotify.clear()
+
+      const yesAction = () => {
+        toast.valid = true // default value
+        this.$snotify.remove(toast.id)
+        this.deleteDesktop(this.desktop.id)
+      }
+
+      const noAction = (toast) => {
+        this.$snotify.remove(toast.id) // default
+      }
+
+      this.$snotify.prompt(`'${this.getCardTitle}' ${i18n.t('messages.confirmation.delete-desktop')}`, `${i18n.t('messages.title.delete')}`, {
+        position: 'centerTop',
+        buttons: [
+          { text: `${i18n.t('messages.yes')}`, action: yesAction, bold: true },
+          { text: `${i18n.t('messages.no')}`, action: noAction }
+        ],
+        placeholder: ''
+      })
     }
   },
   computed: {
@@ -249,8 +275,12 @@ export default {
       MAX_TITLE_SIZE,
       MAX_DESCRIPTION_SIZE,
       MAX_TEMPLATE_TEXT_SIZE,
-      MAX_VIEWER_TEXT_SIZE
+      MAX_VIEWER_TEXT_SIZE,
+      show: false
     }
+  },
+  destroyed () {
+    this.$snotify.clear()
   }
 }
 </script>